+++ Test/Unit/Cron/ExportOmittedAbandonedCartsTest.php
@@ -13,6 +13,7 @@ use CommerceLeague\ActiveCampaign\Test\Unit\AbstractTestCase;
 use CommerceLeague\ActiveCampaign\Model\ResourceModel\Quote\Collection as QuoteCollection;
 use Magento\Framework\MessageQueue\PublisherInterface;
 use PHPUnit\Framework\MockObject\MockObject;
+use Psr\Log\LoggerInterface;
 
 class ExportOmittedAbandonedCartsTest extends AbstractTestCase
 {
@@ -38,6 +39,11 @@ class ExportOmittedAbandonedCartsTest extends AbstractTestCase
     protected $publisher;
 
     /**
+     * @var MockObject|LoggerInterface
+     */
+    protected $logger;
+
+    /**
      * @var PublishOmittedAbandonedCarts
      */
     protected $exportOmittedAbandonedCarts;
@@ -45,6 +51,7 @@ class ExportOmittedAbandonedCartsTest extends AbstractTestCase
     protected function setUp(): void
     {
         $this->configHelper = $this->createMock(ConfigHelper::class);
+        $this->logger = $this->createMock(LoggerInterface::class);
 
         $this->quoteCollectionFactory = $this->getMockBuilder(QuoteCollectionFactory::class)
             ->disableOriginalConstructor()
@@ -62,7 +69,8 @@ class ExportOmittedAbandonedCartsTest extends AbstractTestCase
         $this->exportOmittedAbandonedCarts = new PublishOmittedAbandonedCarts(
             $this->configHelper,
             $this->quoteCollectionFactory,
-            $this->publisher
+            $this->publisher,
+            $this->logger
         );
     }
 
@@ -106,9 +114,12 @@ class ExportOmittedAbandonedCartsTest extends AbstractTestCase
             ->method('isAbandonedCartExportEnabled')
             ->willReturn(true);
 
-        $this->configHelper->expects($this->once())
+        $this->configHelper->expects($this->exactly(2))
             ->method('isRetryAllOmittedEnabled')
             ->willReturn(false);
+
+        $this->logger->expects($this->never())
+            ->method('warning');
 
         $this->quoteCollection->expects($this->once())
             ->method('addAbandonedFilter')
@@ -154,9 +165,13 @@ class ExportOmittedAbandonedCartsTest extends AbstractTestCase
             ->method('isAbandonedCartExportEnabled')
             ->willReturn(true);
 
-        $this->configHelper->expects($this->once())
+        $this->configHelper->expects($this->exactly(2))
             ->method('isRetryAllOmittedEnabled')
             ->willReturn(true);
+
+        $this->logger->expects($this->once())
+            ->method('warning')
+            ->with($this->stringContains('retry_all_omitted is ON'));
 
         $this->quoteCollection->expects($this->never())
             ->method('addAbandonedFilter');