+++ Test/Unit/Cron/ExportOmittedGuestCustomersTest.php
@@ -12,6 +12,7 @@ use CommerceLeague\ActiveCampaign\Model\ResourceModel\ActiveCampaign\GuestCustom
 use CommerceLeague\ActiveCampaign\Test\Unit\AbstractTestCase;
 use Magento\Framework\MessageQueue\PublisherInterface;
 use PHPUnit\Framework\MockObject\MockObject;
+use Psr\Log\LoggerInterface;
 
 class ExportOmittedGuestCustomersTest extends AbstractTestCase
 {
@@ -37,6 +38,11 @@ class ExportOmittedGuestCustomersTest extends AbstractTestCase
     protected $publisher;
 
     /**
+     * @var MockObject|LoggerInterface
+     */
+    protected $logger;
+
+    /**
      * @var PublishOmittedGuestCustomers
      */
     protected $exportOmittedGuestCustomers;
@@ -44,6 +50,7 @@ class ExportOmittedGuestCustomersTest extends AbstractTestCase
     protected function setUp(): void
     {
         $this->configHelper = $this->createMock(ConfigHelper::class);
+        $this->logger = $this->createMock(LoggerInterface::class);
 
         $this->customerCollectionFactory = $this->getMockBuilder(CustomerCollectionFactory::class)
             ->disableOriginalConstructor()
@@ -61,7 +68,8 @@ class ExportOmittedGuestCustomersTest extends AbstractTestCase
         $this->exportOmittedGuestCustomers = new PublishOmittedGuestCustomers(
             $this->configHelper,
             $this->customerCollectionFactory,
-            $this->publisher
+            $this->publisher,
+            $this->logger
         );
     }
 
@@ -103,9 +111,12 @@ class ExportOmittedGuestCustomersTest extends AbstractTestCase
             ->method('isCustomerExportEnabled')
             ->willReturn(true);
 
-        $this->configHelper->expects($this->once())
+        $this->configHelper->expects($this->exactly(2))
             ->method('isRetryAllOmittedEnabled')
             ->willReturn(false);
+
+        $this->logger->expects($this->never())
+            ->method('warning');
 
         $this->customerCollection->expects($this->once())
             ->method('addOmittedFilter')
@@ -143,9 +154,13 @@ class ExportOmittedGuestCustomersTest extends AbstractTestCase
             ->method('isCustomerExportEnabled')
             ->willReturn(true);
 
-        $this->configHelper->expects($this->once())
+        $this->configHelper->expects($this->exactly(2))
             ->method('isRetryAllOmittedEnabled')
             ->willReturn(true);
+
+        $this->logger->expects($this->once())
+            ->method('warning')
+            ->with($this->stringContains('retry_all_omitted is ON'));
 
         $this->customerCollection->expects($this->once())
             ->method('addOmittedFilter')